+++ Tests/Runtime/OpenXRRuntimeTests.cs
@@ -33,7 +33,8 @@ namespace UnityEngine.XR.OpenXR.Tests
                 "XR_MSFT_secondary_view_configuration",
                 "XR_EXT_eye_gaze_interaction",
                 "XR_MSFT_hand_interaction",
-                "XR_MSFT_first_person_observer"
+                "XR_MSFT_first_person_observer",
+                "XR_META_performance_metrics"
             };
 
             foreach (string expectedExtension in expectedExtensions)
@@ -677,6 +678,29 @@ namespace UnityEngine.XR.OpenXR.Tests
             Assert.IsTrue(secondaryLayerCount == 0);
         }
 
+        [UnityTest]
+        public IEnumerator FirstPersonObserverInvalidSecondaryView()
+        {
+            AddExtension("XR_MSFT_secondary_view_configuration");
+            AddExtension("XR_MSFT_first_person_observer");
+            base.InitializeAndStart();
+
+            MockRuntime.ActivateSecondaryView(XrViewConfigurationType.SecondaryMonoFirstPersonObserver, true);
+
+            yield return new WaitForXrFrame(2);
+
+            MockRuntime.GetEndFrameStats(out var primaryLayerCount, out var secondaryLayerCount);
+            Assert.IsTrue(secondaryLayerCount == 1);
+
+            // make mock runtime return invalid state for xrWaitFrame to make sure we don't crash
+            MockRuntime.ActivateSecondaryView(0, false);
+
+            yield return new WaitForXrFrame(2);
+
+            MockRuntime.GetEndFrameStats(out primaryLayerCount, out secondaryLayerCount);
+            Assert.IsTrue(secondaryLayerCount == 0);
+        }
+
 
         [UnityTest]
         public IEnumerator ThirdPersonObserver()
@@ -811,6 +835,12 @@ namespace UnityEngine.XR.OpenXR.Tests
             Assert.IsTrue(OpenXRLoaderBase.Instance == null);
         }
 
+        /// <summary>
+        /// Simulates what can happen when trying to reconnect Link mode after a link disconnect, but
+        /// the headset is reporting a form factor unavailable error.
+        /// In this case, the app has already been started, and the runtime reports a form factor unavailable
+        /// error when trying to initialize xr again. What should happen is a restart loop to restart XR.
+        /// </summary>
         [UnityTest]
         public IEnumerator RestartLoopTest()
         {
@@ -861,6 +891,9 @@ namespace UnityEngine.XR.OpenXR.Tests
             }
         }
 
+        /// <summary>
+        /// Tests that OpenXRRuntime.wantsToRestart is a global switch for de-activating the restarting of XR.
+        /// </summary>
         [UnityTest]
         public IEnumerator RestartLoopDisabledTest()
         {
@@ -903,6 +936,11 @@ namespace UnityEngine.XR.OpenXR.Tests
             }
         }
 
+        /// <summary>
+        /// Simulates what happens when trying to initialize xr for the first time, and the headset is disconnected,
+        /// causing xrGetSystem to report a form factor unavailable error.
+        /// By default, xr initialization should not be retried per feedback from Microsoft.
+        /// </summary>
         [UnityTest]
         public IEnumerator RestartLoopDisabledBeforeInitializationTest()
         {
@@ -936,6 +974,70 @@ namespace UnityEngine.XR.OpenXR.Tests
             }
             finally
             {
+                OpenXRRestarter.TimeBetweenRestartAttempts = initialTimeBetweenRestarts;
+                MockRuntime.SetFunctionCallback("xrGetSystem", initialXRGetSystemCallback);
+                OpenXRLoaderBase.Internal_SetSuccessfullyInitialized(false);
+            }
+        }
+
+        /// <summary>
+        /// Simulates what happens when trying to initialize xr for the first time, and the runtime reports
+        /// a form factor unavailable error for xrGetSystem. If the user chooses, the runtime can retry initialization again and again
+        /// while waiting for the xrGetSystem call to succeed.
+        /// By default, xr initialization should not be retried per feedback from Microsoft. This loop needs to be
+        /// explicitly enabled.
+        /// </summary>
+        [UnityTest]
+        public IEnumerator XRGetSystemLoopBeforeInitializationTest()
+        {
+            OpenXRLoaderBase.Internal_SetSuccessfullyInitialized(false);
+            float initialTimeBetweenRestarts = OpenXRRestarter.TimeBetweenRestartAttempts;
+            bool initialKeepFunctionCallbacks = MockRuntime.KeepFunctionCallbacks;
+            var initialXRGetSystemCallback = MockRuntime.GetBeforeFunctionCallback("xrGetSystem");
+            bool initialRetryInitializationOnFormFactorErrors = OpenXRRuntime.retryInitializationOnFormFactorErrors;
+            try
+            {
+                yield return null;
+
+                // Enable this to prevent xrGetSystem callback override from getting overwritten.
+                MockRuntime.KeepFunctionCallbacks = true;
+
+                // Reduce the time between restarts to reduce the time of this test.
+                OpenXRRestarter.TimeBetweenRestartAttempts = 1.0f;
+
+                // Enable the xrGetSystem retry loop per this test.
+                OpenXRRuntime.retryInitializationOnFormFactorErrors = true;
+
+                // Enable this to ignore the error messages in the logs.
+                LogAssert.ignoreFailingMessages = true;
+
+                int resetAttempts = 0;
+                MockRuntime.SetFunctionCallback("xrGetSystem", (name) =>
+                {
+                    MockRuntime.KeepFunctionCallbacks = true;
+
+                    resetAttempts += 1;
+                    if (resetAttempts <= 2)
+                    {
+                        return XrResult.FormFactorUnavailable;
+                    }
+                    else
+                    {
+                        return XrResult.Success;
+                    }
+                });
+
+                // Trigger initialize, which should call into xrGetSystem.
+                base.InitializeAndStart();
+
+                yield return new WaitForLoaderRestart(10, true);
+                Assert.AreEqual(3, resetAttempts);
+                Assert.IsTrue(OpenXRLoader.Instance != null, "OpenXR should be initialized");
+            }
+            finally
+            {
+                MockRuntime.KeepFunctionCallbacks = initialKeepFunctionCallbacks;
+                OpenXRRuntime.retryInitializationOnFormFactorErrors = initialRetryInitializationOnFormFactorErrors;
                 OpenXRRestarter.TimeBetweenRestartAttempts = initialTimeBetweenRestarts;
                 MockRuntime.SetFunctionCallback("xrGetSystem", initialXRGetSystemCallback);
                 OpenXRLoaderBase.Internal_SetSuccessfullyInitialized(false);
@@ -1053,6 +1155,11 @@ namespace UnityEngine.XR.OpenXR.Tests
             Assert.IsTrue(OpenXRLoader.Instance == null, "OpenXR should not be initialized");
         }
 
+        /// <summary>
+        /// Simulates what can happen when trying to reconnect Link mode after a link disconnect. During
+        /// xr re-initialization, creating the swapchain can result in a session lost error, which should trigger a
+        /// loop to attempt to restart xr.
+        /// </summary>
         [UnityTest]
         public IEnumerator CreateSwapChainSessionLostError()
         {
@@ -1067,7 +1174,6 @@ namespace UnityEngine.XR.OpenXR.Tests
                 OpenXRRestarter.TimeBetweenRestartAttempts = timeBetweenRestarts;
 
                 MockRuntime.SetFunctionCallback("xrCreateSwapchain", (func) => XrResult.SessionLost);
-                LogAssert.Expect(LogType.Log, "OpenXRLoader restart successful.");
                 InitializeAndStart();
 
                 yield return new WaitForLoaderRestart(10, true);